4.0.3

+++ exercicio01.js
@@ -1,4 +1,4 @@
-var celsius = parseInt(prompt("Digite a temperatura em ºC (escreva apenas os números e utilize '.') : "))
+var celsius = parseFloat(prompt("Digite a temperatura em ºC (escreva apenas os números e utilize '.') : "))
 var fahrenheit = ((celsius*1.8)+32)
 var kelvin = (celsius+273.15)
